generalize the editor buttons, move link editor to top left of eval frame
