+++ src/components/full-screen-editor-button.jsx
@@ -1,30 +1,10 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import Tooltip from 'material-ui/Tooltip'
 import EditorIcon from 'material-ui-icons/Code'
 import EvalFrameIcon from 'material-ui-icons/InsertChart'
-import LinkIcon from 'material-ui-icons/Link'
-import SvgIcon from 'material-ui/SvgIcon'
+import EditorButton from './editor-button'
 import tasks from '../actions/task-definitions'
 
-const UnlinkIcon = props => (
-  <SvgIcon {...props}>
-    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
-      <path fill="none" d="M0 0h24v24H0V0z" />
-      <path d="M17 7h-4v1.9h4c1.71 0 3.1 1.39 3.1 3.1 0 1.43-.98 2.63-2.31 2.98l1.46 1.46C20.88 15.61 22 13.95 22 12c0-2.76-2.24-5-5-5zm-1 4h-2.19l2 2H16zM2 4.27l3.11 3.11C3.29 8.12 2 9.91 2 12c0 2.76 2.24 5 5 5h4v-1.9H7c-1.71 0-3.1-1.39-3.1-3.1 0-1.59 1.21-2.9 2.76-3.07L8.73 11H8v2h2.73L13 15.27V17h1.73l4.01 4L20 19.74 3.27 3 2 4.27z" />
-      <path fill="none" d="M0 24V0" />
-    </svg>
-  </SvgIcon>
-)
-
-const EditorButton = props => (
-  <Tooltip classes={{ tooltip: 'iodide-tooltip' }} title={props.text}>
-    <button
-      className={`display-button ${props.isActive ? 'is-active' : ''}`}
-      onClick={props.onClick}
-    >{props.children}
-    </button>
-  </Tooltip>)
 
 export class FullScreenEditorButtonUnconnected extends React.Component {
   render() {
@@ -32,6 +12,9 @@ export class FullScreenEditorButtonUnconnected extends React.Component {
       <div
         className="display-buttons"
         style={{
+        position: 'absolute',
+        right: -25,
+        top: 0,
         display: this.props.viewMode === 'REPORT_VIEW' ? 'none' : undefined,
       }}
       >
@@ -50,13 +33,7 @@ export class FullScreenEditorButtonUnconnected extends React.Component {
         >
           <EvalFrameIcon style={{ fontSize: '13px' }} />
         </EditorButton>
-        <EditorButton
-          text={this.props.scrollingLinked ? 'unscroll w/ editor' : 'scroll w/ editor'}
-          isActive={this.props.scrollingLinked}
-          onClick={() => { tasks.toggleEditorLink.callback(); }}
-        >
-          {this.props.scrollingLinked ? <LinkIcon style={{ fontSize: '13px' }} /> : <UnlinkIcon style={{ fontSize: '13px' }} />}
-        </EditorButton>
+
       </div>
     )
   }
@@ -67,7 +44,6 @@ function mapStateToProps(state) {
     showFrame: state.showFrame,
     showEditor: state.showEditor,
     viewMode: state.viewMode,
-    scrollingLinked: state.scrollingLinked,
   }
 }
