+++ src/components/menu/notebook-header.jsx
@@ -5,7 +5,6 @@ import PresentationModeToolbar from './presentation-mode-toolbar'
 import EditorModeToolbar from './editor-mode-toolbar'
 
 import AppMessages from '../app-messages/app-messages'
-import FullScreenEditorButton from '../full-screen-editor-button'
 
 const theme = createMuiTheme({
   palette: {
@@ -21,7 +20,6 @@ export default class NotebookHeader extends React.Component {
         <MuiThemeProvider theme={theme}>
           <EditorModeToolbar />
         </MuiThemeProvider>
-        <FullScreenEditorButton />
         <PresentationModeToolbar />
         <AppMessages />
       </div>
